Extract static_file_storage crate
Pulls the slot-indexed append-only file format out of static_cold.rs
into a standalone crate at common/static_file_storage. The library
exposes a single StaticFile type that owns one directory of data_*,
data_*.off, and column.conf files; all file-format, durability, and
recovery semantics live there. u64 slots, no EthSpec, no Lighthouse
types - usable by any consumer (HotColdDB cold backend today, lcli
era-import next).

static_cold.rs shrinks 853 -> 192 lines: keeps StaticColdStore<E>,
the DBColumnCold -> ColumnConfig mapping table, the embedded
BeaconNodeBackend<E> for root-keyed indices at <root>/index/, and
the ColdStore<E> impl. Everything else is delegated.

Adds a criterion bench (benches/large_writes.rs) targeting the
performance-critical write path during ERA load: ~30 MB
state-snapshot records into a no-compression column. Four benches:
put_single, put_batch_eight_in_one_file (no-compression and snappy
variants), put_batch_32_cross_file. Smoke run: ~615 MiB/s on
put_single_30mb.

TODO-archive-read-paths.md captures the open research question that
gates any read-path bench or design work: who uses Lighthouse as an
archive and which read endpoints are performance-critical.

Verified: cargo check --workspace --tests clean; 80/80 store_tests
pass under COLD_BACKEND=static.
