+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
     "common/oneshot_broadcast",
     "common/pretty_reqwest_error",
     "common/slot_clock",
+    "common/static_file_storage",
     "common/system_health",
     "common/target_check",
     "common/task_executor",
@@ -217,6 +218,7 @@ signing_method = { path = "validator_client/signing_method" }
 slasher = { path = "slasher", default-features = false }
 slashing_protection = { path = "validator_client/slashing_protection" }
 slot_clock = { path = "common/slot_clock" }
+static_file_storage = { path = "common/static_file_storage" }
 smallvec = "1"
 snap = "1"
 ssz_types = { version = "0.14.0", features = ["context_deserialize", "runtime_types"] }
