+++ beacon_node/store/Cargo.toml
@@ -29,6 +29,7 @@ serde = { workspace = true }
 smallvec = { workspace = true }
 snap = { workspace = true }
 ssz_types = { workspace = true }
+static_file_storage = { workspace = true }
 state_processing = { workspace = true }
 strum = { workspace = true }
 superstruct = { workspace = true }
