+++ common/static_file_storage/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "static_file_storage"
+version = "0.1.0"
+edition = { workspace = true }
+
+[dependencies]
+parking_lot = { workspace = true }
+snap = { workspace = true }
+
+[dev-dependencies]
+criterion = { workspace = true }
+tempfile = { workspace = true }
+
+[[bench]]
+name = "large_writes"
+harness = false
